Lifecycle遷移をint基底のCASへ統一

+++ src/ComputeSharp/Graphics/Resources/Lifetime/ResourceGenerationState.cs
@@ -1,6 +1,6 @@
 namespace ComputeSharp.Resources.Lifetime;
 
-internal enum ResourceGenerationState : byte
+internal enum ResourceGenerationState : int
 {
     Constructing = 0,
     Active = 1,
